+++ CodeMaid.UnitTests/CodeMaidPackageTest.cs
@@ -4,6 +4,8 @@ using SteveCadwallader.CodeMaid.UnitTests.Helpers;
 namespace SteveCadwallader.CodeMaid.UnitTests
 {
     [TestClass]
+    [Ignore]
+    //TODO: Disabled while experimenting with VS2017 RC.
     public class CodeMaidPackageTest
     {
         [TestMethod]